added default product categories in populate_db.js. also added category_ids length check in database.js to prevent doing unnecessary queries

+++ backend/database.js
@@ -183,6 +183,7 @@ class DB {
     }
 
     static async set_product_categories(product_id, category_ids) {
+        if (category_ids.length < 1) return;
         await db.query("DELETE FROM Product_to_categories WHERE product_id = ?", [product_id]);
         for (let category_id of category_ids) {
             await db.query("INSERT INTO Product_to_categories(product_id, category_id) VALUES (?, ?)", [product_id, category_id]);
